Fix some issues caught with deepscan
